Register four new quick-reference shortcodes for quickly linking to function, hook, class, and method references.
The purpose of these shortcodes is to provide a simple, inline method for linking to other reference articles in the Code Reference. Each of the four shortcodes accepts a 'name' and 'label' attribute, and the method shortcode accepts a 'class' attribute for pairing it with the correct class. For each shortcode use, the function reference link HTML is built, then the URL is checked for validity before output.

The shortcodes are registered and output via a new class, WPORG_Shortcodes.

Example usage:
```
[function name="wp_parse_args" label="wp_parse_args()"]
```

```
[hook name="new_status_post-post_type" label="publish_post"]
```

```
[class name="WP_Meta_Query"]
```

```
[method name="have_posts" class="WP_Query" label="WP_Query->have_posts()"]
```